Move useParameterizeTradeInMenu() hook to Updater folder
